fix(logic): fix logic
